[Fix] Fixed getting login token on mfa restore (#464)
* Fixed getting login token on mfa restore

* Fixed ethers import

* Fixing storage for restore
